remove authentication
